feat(GHS): rename secretariat to secretary
