+++ src/app/_components/secretary/webSocketClient.ts
@@ -1,8 +1,8 @@
-import type { GameState } from '@/domain/secretariat/game.state';
+import type { GameState } from '@/domain/secretary/game.state';
 import { setGameState } from '@/stores/game.store';
 
-export function connectToSecretariat(id: string) {
-  const client = new WebSocket('wss://gloomhaven-secretariat.de:8443/');
+export function connectToSecretary(id: string) {
+  const client = new WebSocket('wss://gloomhaven-secretary.de:8443/');
 
   client.onmessage = (event) => {
     const { type, payload } = JSON.parse(event.data) as { type: string, payload: GameState };
@@ -13,7 +13,7 @@ export function connectToSecretariat(id: string) {
 
   client.onopen = () => {
     console.log('connected');
-    localStorage.setItem('secretariat-id', id);
+    localStorage.setItem('secretary-id', id);
     client.send(`{ "code": ${id}, "password": ${id}, "type": "request-game" }`);
   };
