+++ src/domain/secretary/game.state.ts
